segundo commit

+++ src/main/java/br/com/financeiro/modelo/negocio/Lancamento.java
@@ -6,5 +6,5 @@ public interface Lancamento {
 
 	public String getDescricao();
 	public LocalDate getVencimento();
-	public float getValor();
+	public double getValor();
 }
